+++ sampledb/scripts/set_up_demo.py
@@ -4,6 +4,7 @@ Script for setting up demo data in a previously unused SampleDB installation.
 
 Usage: sampledb set_up_demo
 """
+import copy
 import datetime
 import io
 import json
@@ -1429,10 +1430,11 @@ This example shows how Markdown can be used for instrument Notes.
             file_name='test_database.txt',
             save_content=lambda f: typing.cast(None, f.write(b"Test"))
         )
-        eln_file_data = sampledb.logic.export.get_eln_archive(
-            user_id=instrument_responsible_user.id,
-            object_ids=[independent_object.id, file_object.id, measurement.id, user_reference_object.id]
-        )
+        with app.test_request_context('/'):
+            eln_file_data = sampledb.logic.export.get_eln_archive(
+                user_id=instrument_responsible_user.id,
+                object_ids=[independent_object.id, file_object.id, measurement.id, user_reference_object.id]
+            )
         eln_import_id = sampledb.logic.eln_import.create_eln_import(
             user_id=instrument_responsible_user.id,
             file_name='sampledb_demo_file.eln',
@@ -1809,6 +1811,12 @@ This example shows how Markdown can be used for instrument Notes.
                 }
             },
             user_id=instrument_responsible_user.id
+        )
+        horizontal_object_action_schema = copy.deepcopy(typing.cast(typing.Dict[str, typing.Any], horizontal_object_action.schema))
+        horizontal_object_action_schema['displayProperties'] = ['name']
+        sampledb.logic.actions.update_action(
+            action_id=horizontal_object_action.id,
+            schema=horizontal_object_action_schema,
         )
 
     print("Success: set up demo data", flush=True)