+++ tests/logic/schemas/test_data_diffs.py
@@ -3,7 +3,7 @@ import json
 import pytest
 
 import sampledb
-from sampledb.logic.schemas.data_diffs import calculate_diff, apply_diff, VALUE_NOT_SET, invert_diff
+from sampledb.logic.schemas.data_diffs import calculate_diff, apply_diff, VALUE_NOT_SET, invert_diff, iter_diff
 from sampledb.logic.errors import DiffMismatchError
 
 
@@ -738,3 +738,108 @@ def test_array_index_diffs():
         }
     }
     assert invert_diff(data_diff=invert_diff(data_diff=data_diff)) == data_diff
+
+def test_iter_diff():
+    data_before = {
+        "array": [
+            {
+                "_type": "text",
+                "text": "A1"
+            },
+            {
+                "_type": "text",
+                "text": "A2"
+            },
+            {
+                "_type": "text",
+                "text": "A3"
+            }
+        ]
+    }
+    data_after = {
+        "a": {
+            "_type": "text",
+            "text": "1"
+        },
+        "array": [
+            {
+                "_type": "text",
+                "text": "B1"
+            },
+            {
+                "_type": "text",
+                "text": "A2"
+            }
+        ]
+    }
+    assert list(sorted(iter_diff(calculate_diff(
+        data_before=data_before,
+        data_after=data_after
+    )))) == [
+        (
+            ('a',),
+            {
+                "_after": {
+                    "_type": "text",
+                    "text": "1"
+                }
+            }
+        ),
+        (
+            ('array', 0),
+            {
+                "_before": {
+                    "_type": "text",
+                    "text": "A1"
+                },
+                "_after": {
+                    "_type": "text",
+                    "text": "B1"
+                }
+            }
+        ),
+        (
+            ('array', 2),
+            {
+                "_before": {
+                    "_type": "text",
+                    "text": "A3"
+                }
+            }
+        )
+    ]
+
+
+def test_apply_placeholder_diff():
+    schema_before = {
+        "type": "object",
+        "properties": {
+            "text": {
+                "type": "text",
+                "title": {"en": "Text"}
+            }
+        },
+        "required": ["text"]
+    }
+    data_before = sampledb.logic.schemas.generate_placeholder(schema_before)
+    data_after = {
+        "_type": "text",
+        "text": "A"
+    }
+    data_diff = {
+        '_before': data_before,
+        '_after': data_after
+    }
+    assert apply_diff(
+        data_before=data_before,
+        data_diff=data_diff,
+        schema_before=schema_before,
+        validate_data_before=False
+    ) == data_after
+    with pytest.raises((sampledb.logic.errors.DiffMismatchError, sampledb.logic.errors.ValidationError)):
+        apply_diff(
+            data_before=data_before,
+            data_diff=data_diff,
+            schema_before=schema_before,
+            validate_data_before=True
+        )